Fix the connected tests (disable tests which where skipped before by FlakyTest flag)
