+++ synchronization/src/androidTest/AndroidManifest.xml
@@ -12,5 +12,30 @@
             android:process=":persistence_process"
             android:syncable="true"
             tools:replace="android:authorities"/>
+
+        <!-- Some connected synchronization Android tests need this for the test sync to start -->
+        <service
+            android:name="de.cyface.synchronization.CyfaceSyncService"
+            android:exported="false"
+            android:process=":sync">
+            <intent-filter>
+                <action android:name="android.content.SyncAdapter" />
+            </intent-filter>
+            <meta-data
+                android:name="android.content.SyncAdapter"
+                android:resource="@xml/sync_adapter" />
+        </service>
+        <!-- Some connected Android tests need this to create a test account, or else we get -->
+        <!-- SecurityException: ... cannot explicitly add accounts of type: ... -->
+        <service android:name="de.cyface.synchronization.CyfaceAuthenticatorService"
+            android:exported="false"
+            android:process=":sync"> <!-- should be fine to use the sync process -->
+            <intent-filter>
+                <action android:name="android.accounts.AccountAuthenticator" />
+            </intent-filter>
+            <meta-data
+                android:name="android.accounts.AccountAuthenticator"
+                android:resource="@xml/authenticator" />
+        </service>
     </application>
 </manifest>
